02_Chat_Models/requirements.txt
langchain==0.1.9
openai==1.12.0
langchain-openai==0.0.8
streamlit
python-dotenv